Use getCookieCategory rather than assuming category
The global_bar_seen cookie is in the "settings" category. For a quick fix, we previously hardcoded this assumption into Static, which is bad if we decided to change the category in the govuk_publishing_components gem.

We now have a function in the gem where we can provide a cookie name and the category is returned. This commit replaces our hardcoded 'settings' category with that method.