Set initial global bar cookie with helper function
We had an issue where the global bar was setting 2 global_bar_seen cookies on some pages. The initialise logic was setting a cookie with the path of the page, whereas other logic was using GOVUK.setCookie() which sets with a path of "/".

This commit switches the initialise logic to use GOVUK.setCookie() too, so the path is the same and we only have one cookie.