Add a custom meta tag for dimension 21
We will be looking to track whether displayed search results were
as a result of users selecting a suggestion from the title autocomplete
or whether they just used the keyword search.

Dimension 21 was chosen in
https://trello.com/c/FbRA8Wf1/1337-define-tracking-on-title-suggest-autocomplete
and agreed with the Performance Analyst community as per below

"Dimension 21 has been changed on Google analytics GOV.UK main profile.
This was previously "search position" but was no longer being used.
It has been renamed to "Search autocomplete status".
The name can be changed at a later date."