Now matches the exact path for url blocking
Previously it was checking for segments so if we had `/coranavirus` blocked, `/guidance/coronavirus-covid-19-information-for-the-public` would be blocked too as it has `/coronavirus` in the middle.

This change now forces the regex to check exact match so it would only match if the path is `/coronavirus` not `/guidance/coronavirus-covid-19-information-for-the-public`.

Trello: https://trello.com/c/a83JfZDG/60-check-why-the-banner-doesnt-appear-on-certain-pages