Repurpose existing cookie for partial dismissal
The current global_bar_seen cookie, when set to 999, dismisses the whole banner so the user won't see it again.
We want to repurpose this so that instead of hiding the whole banner, it only hides part of it.
The top section (with a link to the coronavirus landing page) will not be dismissable