Fix tests for asset location changing
This makes all the template tests be independent of the prefix in Rails
configuration by using the ends with attribute selector. I'm pretty
baffled why we need to test that we output these specific assets, but I
didn't want to dig too much into this as I think the whole test files
are a bit unusual in what they condition on. I guess they're mostly just
a few assertions to indicate the correct item rendered.

This also updates the assets path for redirects to have the new assets
prefix.