Remove GOVUKTracker
I came across this code while exploring the options for moving static's
asset files from being served from the www host rather than the assets
one. It proved to be a blocking issue as it required the presence of a
"/static/a" on the assets host.

Further investigation led me to question whether this data is actually
used for anything and I could find no evidence that we do anything other
than record this data nor anyone who knew about it. For reference this has
existed for 3 years, stores ~100 GB in S3 and produces HTTP requests that
mirror each Google Analytics one.

The simplest option to get past the blocking issue seemed to be removing
this code, which then allows cleaning up the configuration in AWS and
Fastly CDN. If we decide we have a use for this later we can always
resurrect it.