Add missing pages for known errors
Previously the use of Slimmer meant that all unexpected responses
were treated as '500 Internal Server Error', which is incorrect and
misleading, since the fault is often with the client. This adds new
error pages to cover all public error responses in the last 30 days.

Using our logs, we found the following unhandled errors:

  - 499 is unnecessary, as the client closed the request
  - 405 can be similar to our current 403
  - 401 can be similar to our current 403
  - 429 is for rate limiting (new page)
  - 409 is unnecessary, as it's internal only (conflict)
  - 426 is only handled at the Nginx level