Move showing status code out of the title
Previously we appended the status code onto the end of the browser
title for some of the static error pages, which was inconsistent,
and also restricted us from improving the content of these titles.
This moves the status code into a visually separate area of the
screen, so it can be quoted for reference.