Make 5xx errors consistent with 4xx errors
Previously we defined content for 4xx errors explicitly in each
template, and the content for 5xx errors using conditionals in the
'error_page' partial, which was confusing. This duplicates the 5xx
content into each error template, noting that the duplication is
incidental, and making this change allows us to iterate it.