Iterate design and copy for error pages
This updates the copy on the most of the error pages, following a
content review. We also make a few other changes:

- Move and re-style the "Status code" message
- Synchronise the browser title with the page heading

Previously we had a mixture of browser titles being in-sync (or not)
with the page heading, for no apparent reason. The convention on
GOV.UK is generally to use the same text for both.