Remove asset symlinking without digest
This was put in as a temporary fix a year ago after we found
that there were some resources linking to extremely old assets that
didn't have a digest.

At the point of time now we've just swapped over the hostname for
serving assets which will have broken any remaining links to these
files.

Since this was only a temporary measure in the first place we should be
safe to get rid of this, and if we're not we should really fix whatever
is still calling these rather than redirecting them to another host.