Wrap humans.txt at 100 characters
Long lines of text are hard for people to read. Because humans.txt is
plain text, we can't use the usual HTML / CSS techniques to wrap the
lines, so they have to be wrapped by hand.

Pro tip for vim users:

Set the text width to 100 characters:

`:set tw=100`

Reflow the current paragraph:

`gqap`