Fix search label cut off text
The global site header search label text has been updated from "Search"
to "Search on GOV.UK".
This has lead to an issue where the label text is now appearing clipped
on larger mobile adaptations as there simply is not enough space to
accommodate it.
This updates the CSS so that the search field is hidden on larger
mobiles – same as it currently is for smaller mobiles. The search form
can be shown/hidden using the search toggle button.