Allow stripping query string parameters from Analytics
This adds a new meta tag approach to stripping PII data from being sent
to analytics. This allows for particular query string parameters to be
redacted and for an application to specify them.

To use this an application should create a meta element of:

```
<meta name="govuk:static-analytics:strip-query-string-parameters"
content="param-1, param-2">
```

Where param-1, param-2 is a comma separated list of parameter names.
Formatting for this is inspired by `<meta name=keywords>`

Then when analytics data is sent to GA the strings are looked for a
match of the query string parameter and any found are replaced. To illustrate:
were a url of `https://www.gov.uk?param-1=secret` be sent to analytics,
and param-1 was flagged for stripping, then this was be replaced to be
`https://www.gov.uk?param-1=[param-1]`.

There are a few caveats of this approach that should be outlined:

It's just a regular expression that tries to match query strings and this
is done rather crudely. It just looks for a part of a string which is
`?parameter_name=<some value>` or `&paramater_name=<some value>`, so it is
possible that for some strings this could be a false positive. A more
thorough approach could be to parse URIs that are recognised in input
and break them down to constituent parts to identify query strings, but
it isn't clear this is needed.

This won't match a URL that is urlencoded - for example &amp; - this
feels like something that would occur if we were showing someone
publicly the URL they used - which seems a bit strange for PII.

As the means to specify query string parameters is a comma separated list,
this technique will not support any query string that use a comma in the
parameter name. This seemed pretty low risk to me, were we to want to support
this we may want to switch to a JSON encoded approach to defining parameter
names to strip.