Include yarn lint in default rake task
This creates a lint.rake file which can do Ruby, JS and Sass linting. It
then updates the Rakefile to make use of it (with a quick modernisation
of the load_tasks call).

It also updates the Jenkinsfile job to stop overriding the test task and
instead use the default rake.