static.css → application.css
This application uses a non-standard name for the main scss file. We change this to the standard name and make an alias for the print style sheet. The alias is needed as the component guide requires application.scss and print.scss to exists, whereas the way static layout is currently structured, it requires an application-print.scss. Further clean-up will be possible when moving to a new layout.