Update header search button text
The default text for the button in the sarch component is "Search" - this means
that multiple forms can have the same submit text. This can cause confusion
when navigating out of context - for example, a screenreader will pull out all
of the interactive elements on a page and list them without the surrounding
information.

To avoid this confusion, the button text has been changed from "Search" to the
more specific "Search GOV.UK".