Adds dropdown menus
Adds dropdown menu to topic, government activity and search.

This change includes all of the JS, CSS and HTML needed to render dropdown menus.

Dropdown menus will allow users to navigate through GOVUK more effectively.

Please note:

The designs require the dropdown background to be full width, however, due to accessibilty reasons
we need to have the tab order to be:
1. Navigation link
2. Dropdown content
3. Next Navigation link

Due to the navigation items being within a width-container, the background only extends up to the container width.
To achieve the styling requirements as well as meet accessibility standards - the solution that I have proposed is that we have a
a background that sits outside of the width container and resizes based on the size of the dropdown.