Add account pages to global bar BlockList
The Accounts team are moving the account homepage to the frontend app,
with the view of transitioning other parts of the account manager to
be rendered by frontend, as we retire the
govuk-account-manager-prototype app. Seeing as the frontend app uses
static, the global bar (currently Coronavirus bar) is automatically
added to the new /account/home page. The design for the account
manager excludes this additional bar currently.
This adds paths beginning with /account/ (assuming that any future
account-related pages will live under /account) to the urlBlockList
in the global bar script.