Add gem layout template variation for accounts
The Accounts team are relocating some of the account pages to the
frontend app. Since frontend app uses static, and certain aspects of the
account manager template differ greatly from the default gem layout,
this warrants the introduction of a new template for use on Accounts
pages.
Some main differences are that the Account logo link is for the account
homepage (as opposed to govuk homepage). The Account product name should
be present in the header. The navigation should be comprised of
account-related links.
Additionally, the feedback component at the bottom of the page should be
hidden as the account has its own version of a feedback component/link.