Update Technical Documentation section
This section lacked context on how Static & Slimmer work together,
and the valuable insight that Static & Slimmer are intended to be
removed in the long term.

It soon became clear we have two kinds of documentation: knowledge,
and "How to's". I've created a "How to's" subsection that follows
the same grammatical conventions as established in
https://github.com/alphagov/govuk-docker/blob/main/docs/how-tos.md