Delete startup.sh
There is no reason for startup.sh to exist, as unlike
our other frontend apps which have a use for startup.sh being used
against live environments, Static is a templating application and
there's no use case we can think of where you'd need to point to
live.

I've therefore removed references to `startup.sh` in the README. I
did consider adding the `govuk-docker-run` prefix to the commands
in the README, as we no longer need to make the commands apply to
both govuk-docker and startup.sh, but this couples us to the
govuk-docker API, which could change in the future. Better to link
to external README.