Add account pages to global survey blocklist
Disable the survey prompt for pages under /account. The reason for this
is that there are already other prompts for feedback on all account
pages. Having yet another one adds unnecessary noise. Another reason is
to keep gov.uk/account/home (rendered by frontend) consistent with the
other account pages which are currently still rendered by the
govuk-account-manager-prototype, and do not use the user satisfaction survey.