Add new layout template for account-like pages
This adds a new account-related template, intended to be used on pages
that should have the account layout, while not being rendered by the
govuk-account-manager-prototype app.

An example of such a page is the gov.uk/account/home page which was
relocated from the account manager app to the frontend app, which at
that time meant that some of the content and page furniture was
duplicated between those applications.

More recently, the accounts team are working towards the goal of single
page notifications which will enable users who have linked their email
notifications address and their GOVUK Account to manage their email
subscriptions through their account.

Part of this work is making the email management pages (which live in
email-alert-frontend) look the same as the account manager when the user
is authenticated with a linked account.

Thanks to the changes included in this commit, we should be able to
achieve this by simply requesting the gem_layout_account_manager from
static, instead of copying individual components to yet another app.