Account layout template adjustments
Implement some adjustments to the layout template for accounts to ensure
a smooth transition to DI auth.

Remove the Account product name from the header, and ensure the GOVUK
link leads back to the GOVUK homepage so users don't get stuck in the
account zone.

Remove footer navigation links in favour or the more stripped down
version of the footer. We want the footer to match what DI have
implemented (it could look weird if the footer dramatically changes when
someone navigates from one account page to another)