Add account no nav template variation
With the migration to DI, we are moving some pages from the account
manager prototype app to `frontend`. Some of these interstitial pages
(such as the "Control how we use information about you page") should
have the account elements, such as the account header and footer and
account phase banner, but should not include the account side navigation
component.
The gem_layout_account_manager_no_nav template has been created with
this purpose.