Clear the template caches every 5 minutes
Rather than requiring a separate script run by Jenkins [1]
to clear the caches, Static can clear it's own cache using
the same ActionPack page caching API:
https://github.com/rails/actionpack-page_caching

This ensures that this process is tested (as compared to the
Jenkins script, which is hard to test).

It also ensures that the template caches expire - this design
is similar to the 'sweepers' recommended in the ActionPack page
cache docs.

Longer term we probably want to look at replacing Static or at
least making the serving less dynamic - there's no need for
a Ruby webserver to be involved since it's Nginx that mostly
is serving these files.

[1]: https://github.com/alphagov/govuk-puppet/blob/178ff3c0d1b024a5e0c81f384b07d0dfb65f418d/modules/govuk_jenkins/templates/jobs/clear_template_cache.yaml.erb#L18