Account layout for "manage your account" active
When the account layout template renders with a nav, the "Your account"
nav option is selected by default.

Some pages which use the account layout belong under "Manage your
account" instead of "Your account".

This allows us the option to render the account layout with "Manage your
account" selected/active in the account nav, instead of the default
"Your account".