Add tests for Coronavirus link removal What
Now that the Coronvirus link has been removed from all 4xx and 5xx pages, we
should test to make sure that it is not longer visible.

Unfortunately, a test for the existence of the Coronavirus link on these
pages was not previously added.

Why

We are removing all signposting to Coronavirus across GOV.UK.