Remove emergency banner component
As part of the work to retire static, the emergency
banner component is being moved to govuk_publishing_components.
This will mean that there will be no components/layouts left
in static and so therefore frontend apps will not be dependent on
static for any styles or components.

The partials, styles and docs removed in this commit have been
moved to the gem.