Replace calls to partial with calls to gem calls
Emergency banner has been moved to govuk_publishing_components. All
calls to the emergency_banner within static have been replaced to
calls to emergency_banner within the gem. Imports to stylesheets
of emergency_banner have been added to application.scss and
application-print.scss.

Have moved the default link text behaviour to the gem, so it has
been removed from static in this commit.

Updated the tests to expect the new component instead of placeholder
HTML that was used to test the previous implementation of emergency
banner.