Add gem_base_homepage template
Emergency banner was moved to govuk_publishing_components which means
it had to be refactored to meet the component standards. Part of this
was removing a parent selector in the css which would apply
alternative styles to the emergency banner if the page the
banner appears on is the homepage.

To remove this external selector, the emergency banner now
accepts a 'homepage' param which will add modifiers to classes
if this param is set. We therefore need a new gem_layout template
specifically for homepage, where this param is set to true so
we can style the emergency banner differently on homepage. This
gem layout will need to be updated in frontend to work.