New template with no emergency banner
When the emergency banner is deployed to GOV.UK and it links to a page on
GOV.UK, we do not want the banner to appear on the page itself. If this
new template (gem_layout_no_emergency_banner) is specified then the
emergency banner will not be rendered on that page.