Spike: show a banner warning when page is rendered via mirror
If the user views a mirrored page, there is no indication that it
is a mirrored page (except by accident, e.g. looking broken). The
page content may be out of date, depending on when it was mirrored.
Additionally, dynamic content won't work, so if Origin is down and
we're serving mirrored content, if the user performs a "Search",
for example, they'll get a 503 error.

I've used the [Notice component][] as there is precedent for using
it to indicate content warnings, e.g.
https://components.publishing.service.gov.uk/component-guide/notice/with_locale.

The wording and styling are of course subject to change.

Trello: https://trello.com/c/OuKvzs4L/2741-write-rfc-to-be-transparent-about-mirrored-content

[Notice component]: https://components.publishing.service.gov.uk/component-guide/notice