Add new template for blue bar full width
This adds a new template that allows a blue bar to be added to full
width templates. Adding this template avoids having to implement the bar
within the collections application itself, which was causing styling
differences to the blue bar that appears on regular width layouts when
browsing collections pages.