Update dev config for asset proxying
This adjusts this application so that the static asset proxying [1] can
be used as the default approach for handling assets in this application.

The motivation for this is an upcoming change in the GOV.UK Content
Security Policy which restricts the hosts used for assets [2]. This
change offers an opportunity to make the app behave more consistently
with production.

[1]: https://github.com/alphagov/govuk_app_config/blob/7f060692720df50a27f6845f052b04eae2246226/lib/govuk_app_config/govuk_proxy/static_proxy.rb
[2]: https://github.com/alphagov/govuk_app_config/pull/274