Remove production ASSET_HOST configuration
The ASSET_HOST env var is no longer used in production and the Heroku
one has likely been defunct for years as it falls foul of the Content
Security Policy.

The ASSET_HOST value has not been used since the introduction of of the
Static proxy [1] which has allowed a relative path for static assets.

This has been removed because of an upcoming change to Content Security
Policy which would cause setting this to fall foul of the CSP [2].

If apps wish to utilise a static that is hosted on Heroku then they
should specify a PLEK_SERVICE_STATIC_URI env var of the Heroku url and ensure
GOVUK_PROXY_STATIC_ENABLED=true. This would work despite the CSP.

[1]: https://github.com/alphagov/govuk_app_config/blob/7f060692720df50a27f6845f052b04eae2246226/lib/govuk_app_config/govuk_proxy/static_proxy.rb
[2]: https://github.com/alphagov/govuk_app_config/pull/274