Raise exception if unexpected arguments passed to rake task
It's possible to accidentally pass more arguments than expected
to the rake task; if one of the arguments contains an unescaped
`,`, then the input will split at that point and be treated as
two inputs. This is all too easy to do, and the consequences are
surprisingly bad: every input is out of sync with the expected
argument, and logic somewhere in Origin gets confused and ultimately
starts serving 500 errors!

This commit does a sense-check on the inputs, to ensure that no
more than 5 arguments are passed. This gives some sense of
security, but isn't perfect, as the task only has 2 mandatory
inputs and the other 3 are optional, so some further validation
will be added in the next commit.

Trello: https://trello.com/c/ocrS67MT/3222-make-emergency-banner-rake-task-more-robust