Upgrade Ruby version
Upgrades `.ruby-version` file which is used in Github CI and local
development.

Replaces the specified `ruby` version in the Gemfile from
`.ruby-version` file to a floating patch Ruby version (3.2.x). This has
the benefit of allowing patch versions of Ruby to be deployed by
updating the Ruby Docker image tag and rebuilding the app.

Trello:
https://trello.com/c/nKilDK4W/2198-dart-sass-upgrade-ruby-to-314-across-govuk-publishing-components-and-other-ux-apps