Improve data-ga4-global-bar attribute inclusion
When the user is on the same page that the global bar links to, the global bar is hidden via CSS. However the HTML still exists, so our data-ga4-global-bar attribute was in the DOM even though the global bar was not visible. This led to incorrect tracking on the page that the global bar was linking to, as the global bar was tracked as being shown to the user when it was not.