Update icon_redirects_test.rb
The test will now only check that the favicon.ico file is redirected.

Uncommented a test that was previously broken when checking for apple touch icons. It appears that in order for the tests to pass, there needs to be an entry in both routes.rb and publishing_api.rake, for example favicon.ico and apple-touch-icon.png pass (referenced in both files), but apple-touch-icon-152x152 does not (only referenced in routes.rb)