Default to EMERGENCY_BANNER_REDIS_URL in Banner's Redis
This is a new environment variable we've currently added to Integration
to allow us to using Whitehall's Redis instance, which is shared with
Static.

For environments that don't have this variable set yet (i.e. Staging and
Production) we fallback on the default `REDIS_URL`.

This is temporary - once we've moved all environments over to the new
URL we can remove this default.

`mock_env` is shamelessly copied from
[here](https://gist.github.com/jazzytomato/79bb6ff516d93486df4e14169f4426af)
and allows us to specify some environment variables that won't persist
between tests.

The memoization of the instance variable in the class method was a
tricky one to overcome, as the setup persisted across tests. I've had to
manually clear it on these new tests to get this working.