Use only `EMERGENCY_BANNER_REDIS_URL` environment variable
In 9683ede220f184f188ff4b86e68461b2f09b2872, we allowed this application
to use either `EMERGENCY_BANNER_REDIS_URL` or `REDIS_URL`, as different
environments were in differet states.

That is no longer the case, so we can remove the code to default to one
over the other.