Add search-with-autocomplete component as assets
This adds in search-with-autocomplete to the global assets that static
uses, this is added because the layout-super-navigation-header makes use
of the search-with-autocomplete component.

The motivation for this change was to resolve a bug when we pushed
search-with-autocomplete live and we discovered that the autocomplete
only rendered and operated correctly on a few pages that already had
the component embedded. The reason this was not operational was because
the CSS and JavaScript were not available unless the autocomplete was
already embedded.

There was a decision to make as to whether to add the dependency to the
layout-super-navigation-header component or to add them separately here.
I've gone with the separate approach as that was the precedent already
used for the search nested component.

Following this change I'm going to go through and remove the previous
embeddings of the JS and CSS so that we don't have it included twice.